WebApp/models/cnndetection_image/requirements.txt
numpy==1.21.2
Pillow==9.0.1
torch==1.10.2
torchvision==0.11.3
